Audio path bug fix.
I incorrectly was checking the audio source against a ``file://`` to see
if it was local or not. The custom formatting string I was using was not
handling non-English characters. It turns out the HTML5 audio doesn't
need a local path formatted this way. Using ``path.normalize`` also help
ensure success.